+++ src/nwbinspector/checks/ecephys.py
@@ -3,8 +3,9 @@ import numpy as np
 
 from pynwb.misc import Units
 from pynwb.ecephys import ElectricalSeries
+from pynwb.file import ElectrodeTable
 
-from hdmf.utils import get_data_shape
+from ..utils import get_data_shape
 
 from ..register_checks import register_check, Importance, InspectorMessage