+++ src/nwbinspector/utils.py
@@ -87,7 +87,7 @@ def is_regular_series(series: np.ndarray, tolerance_decimals: int = 9):
     return len(uniq_diff_ts) == 1
 
 
-def is_ascending_series(series: Union[h5py.Dataset, ArrayLike], nelems=None):
+def is_ascending_series(series: Union[h5py.Dataset, ArrayLike], nelems: Optional[int] = None):
     """General purpose function for determining if a series is monotonic increasing."""
     if isinstance(series, h5py.Dataset):
         return np.all(np.diff(_cache_data_selection(data=series, selection=slice(nelems))) > 0)
@@ -192,3 +192,42 @@ def calculate_number_of_cpu(requested_cpu: int = 1) -> int:
         return requested_cpu
     else:
         return total_cpu + requested_cpu
+
+
+def get_data_shape(data, strict_no_data_load=False):
+    """
+    modified from hdmf.utils.get_data_shape to return shape instead of maxshape
+    Helper function used to determine the shape of the given array.
+
+    In order to determine the shape of nested tuples, lists, and sets, this function
+    recursively inspects elements along the dimensions, assuming that the data has a regular,
+    rectangular shape. In the case of out-of-core iterators, this means that the first item
+    along each dimension would potentially be loaded into memory. Set strict_no_data_load=True
+    to enforce that this does not happen, at the cost that we may not be able to determine
+    the shape of the array.
+
+    :param data: Array for which we should determine the shape.
+    :type data: List, numpy.ndarray, DataChunkIterator, any object that support __len__ or .shape.
+    :param strict_no_data_load: If True and data is an out-of-core iterator, None may be returned. If False (default),
+                                the first element of data may be loaded into memory.
+    :return: Tuple of ints indicating the size of known dimensions. Dimensions for which the size is unknown
+             will be set to None.
+    """
+
+    def __get_shape_helper(local_data):
+        shape = list()
+        if hasattr(local_data, "__len__"):
+            shape.append(len(local_data))
+            if len(local_data):
+                el = next(iter(local_data))
+                if not isinstance(el, (str, bytes)):
+                    shape.extend(__get_shape_helper(el))
+        return tuple(shape)
+
+    if hasattr(data, "shape") and data.shape is not None:
+        return data.shape
+    if isinstance(data, dict):
+        return
+    if hasattr(data, "__len__") and not isinstance(data, (str, bytes)):
+        if not strict_no_data_load or isinstance(data, (list, tuple, set)):
+            return __get_shape_helper(data)