+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Upcoming
 
+### New Checks
+* Add check for unique ids for DynamicTables. [PR #316](https://github.com/NeurodataWithoutBorders/nwbinspector/pull/316)
+
+
+### Fixes
+* Fix `check_subject_proper_age_range` to parse years. [PR #314](https://github.com/NeurodataWithoutBorders/nwbinspector/pull/314)
+* Write a custom `get_data_shape` method that does not return `maxshape`, which fixes errors in parsing shape. [PR #315](https://github.com/NeurodataWithoutBorders/nwbinspector/pull/315)
+
+
 # v0.4.20
 
 ### Improvements
@@ -14,7 +23,6 @@
 ### Fixes
 
 * Fixed issue in `run_checks` following [PR #303](https://github.com/NeurodataWithoutBorders/nwbinspector/pull/303) that prevented iteration over certain check output types. [PR #306](https://github.com/NeurodataWithoutBorders/nwbinspector/pull/306)
-
 
 # v0.4.19